Apache Airflow Projects/Caravan Insurance Challenge/requirements.txt
psycopg2-binary
pandas